refactor(frontend): clearPlaylistCacheをfeature/playlistに移動 #187
